+++ lib/src/addon/legend/legend.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
-import 'package:vector_map/src/data/map_layer.dart';
 import 'package:vector_map/src/addon/map_addon.dart';
+import 'package:vector_map/src/data/map_layer.dart';
 
 /// Abstract legend
 abstract class Legend extends MapAddon {