+++ lib/src/data/map_feature.dart
@@ -35,4 +35,12 @@ class MapFeature {
     }
     return null;
   }
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is MapFeature && runtimeType == other.runtimeType && id == other.id;
+
+  @override
+  int get hashCode => id.hashCode;
 }