+++ lib/src/drawable/drawable_layer.dart
@@ -5,7 +5,8 @@ import 'package:vector_map/src/data/map_data_source.dart';
 import 'package:vector_map/src/data/map_feature.dart';
 import 'package:vector_map/src/data/map_layer.dart';
 import 'package:vector_map/src/drawable/drawable_feature.dart';
-import 'package:vector_map/src/highlight_rule.dart';
+import 'package:vector_map/src/error.dart';
+import 'package:vector_map/src/map_highlight.dart';
 import 'package:vector_map/src/theme/map_theme.dart';
 
 /// Holds all geometry layers to be paint in the current resolution.
@@ -15,6 +16,15 @@ class DrawableLayer {
   final MapLayer layer;
   final Map<int, DrawableFeature> drawableFeatures;
 
+  /// Gets a [DrawableFeature] for a given [MapFeature].
+  DrawableFeature getDrawableFeature(MapFeature feature) {
+    int featureId = feature.id;
+    if (drawableFeatures.containsKey(featureId) == false) {
+      throw VectorMapError('No DrawableFeature for feature id: $featureId');
+    }
+    return drawableFeatures[featureId]!;
+  }
+
   /// Draws the features on a given canvas.
   ///
   /// Only features that match [highlightRule] will be drawn.
@@ -23,7 +33,7 @@ class DrawableLayer {
       required double contourThickness,
       required double scale,
       required bool antiAlias,
-      HighlightRule? highlightRule}) {
+      MapHighlight? highlightRule}) {
     MapDataSource dataSource = layer.dataSource;
     MapTheme theme = layer.theme;
     Color? highlightColor = layer.highlightTheme?.color;
@@ -71,7 +81,7 @@ class DrawableLayer {
       required double contourThickness,
       required double scale,
       required bool antiAlias,
-      HighlightRule? highlightRule}) {
+      MapHighlight? highlightRule}) {
     MapTheme theme = layer.theme;
 
     late Color contourColor;
@@ -99,6 +109,28 @@ class DrawableLayer {
           }
         }
         drawableFeature.drawOn(canvas, paint, scale);
+      }
+    }
+  }
+
+  /// Draws the features that match [MapMultiHighlight] on a given canvas.
+  drawHighlightOn(
+      {required Canvas canvas,
+      required Paint paint,
+      required double scale,
+      required bool fillOnly,
+      required MapHighlight highlight}) {
+    MapDataSource dataSource = layer.dataSource;
+    for (int id in drawableFeatures.keys) {
+      MapFeature feature = dataSource.features[id]!;
+      DrawableFeature drawableFeature = getDrawableFeature(feature);
+      if (drawableFeature.visible) {
+        if (fillOnly && drawableFeature.hasFill == false) {
+          continue;
+        }
+        if (highlight.applies(feature)) {
+          drawableFeature.drawOn(canvas, paint, scale);
+        }
       }
     }
   }